change fontSize from int to double

+++ lib/globals/typography.dart
@@ -2,6 +2,6 @@ class Typography{
   static String latinFontPrimary = "TitilliumWeb";
   static String latinFontSecondary = "Roboto";
 
-  static int paragraphSize = 16;
-  static int focusSize = 23;
+  static double paragraphSize = 16;
+  static double focusSize = 23;
 }
